refactor(api): address CodeRabbit review for sklearn estimators
Improve sklearn estimator type safety and validation:

**Type Safety**
- Fix objective function typing: Callable[[Array], float] -> Callable[[Array], Array]
- Update fit() and score() signatures in RiemannianOptimizer
- JAX grad expects scalar Array output, not Python float

**Validation**
- Add n_components bounds checking in RiemannianPCA.fit()
- Validate 1 <= n_components <= D (tangent dimension)
- Raise ValueError with clear message on invalid n_components

**Test Improvements**
- Match dtype in identity matrix assertions for Gram matrices
- Use weights.dtype for jnp.eye() to avoid dtype mismatches

Addresses CodeRabbit feedback on type annotations and parameter validation.

+++ tests/api/test_flax_nnx.py
@@ -140,7 +140,7 @@ class TestManifoldConstrainedLinear:
 
         # Assert - W^T W should be identity for orthogonal matrices
         gram = weights.T @ weights
-        expected = jnp.eye(out_features)
+        expected = jnp.eye(out_features, dtype=weights.dtype)
         assert jnp.allclose(gram, expected, atol=1e-5)
 
     def test_linear_layer_tracks_constraint_violations(self):
@@ -287,7 +287,7 @@ class TestTrainingWithConstraints:
         # Assert - Should still be orthogonal
         weights = layer.weight.value
         gram = weights.T @ weights
-        expected = jnp.eye(out_features)
+        expected = jnp.eye(out_features, dtype=weights.dtype)
         assert jnp.allclose(gram, expected, atol=1e-4)
 
     def test_jit_compilation_with_constraints(self):
